# public/openapi.yaml
openapi: 3.1.0
info:
  title: "What's on? API"
  description: |
    Aggregate movie and TV show ratings from IMDb, Rotten Tomatoes, AlloCiné,
    Metacritic, Letterboxd, BetaSeries, SensCritique, TMDB, and Trakt.

    **Rate limiting** — 50 requests per hour and 200 per day. Pass an optional
    `api_key` query parameter for a higher limit. The `Retry-After` header is set
    when the limit is reached (HTTP 429).

    **Ratings scale** — `ratings_average` and `minimum_ratings` use a normalised
    0–5 scale across all sources (IMDb ÷ 2, Metacritic ÷ 20, Rotten Tomatoes ÷ 20,
    AlloCiné native, etc.).
  version: 3.39.5
  contact:
    name: Pierre Vanoverschelde
    url: https://github.com/pierrevano/whatson-api
  license:
    name: See LICENSE
    url: https://github.com/pierrevano/whatson-api/blob/main/LICENSE
servers:
  - url: https://whatson-api.onrender.com
    description: Production
security:
  - apiKey: []

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: api_key
      description: Optional API key. Requests without a key are subject to the public rate limit.
  parameters:
...
    releaseDateRange:
      name: release_date
      in: query
      description: |
        Date range filter. Either bound is optional.
        Format: `from:YYYY-MM-DD,to:YYYY-MM-DD`
      schema:
        type: string
        example: "from:2020-01-01,to:2024-12-31"
    genres:
      name: genres
      in: query
      description: Comma-separated genres to filter by (case-insensitive partial match). Pass `all` to skip this filter.
      schema:
        type: string
        example: "Action,Drama"
    platforms:
      name: platforms
      in: query
      description: Comma-separated streaming platform names (case-insensitive partial match). Pass `all` to skip this filter.
      schema:
        type: string
        example: "Netflix,Disney+"
    networks:
      name: networks
      in: query
      description: Comma-separated TV network names (case-insensitive partial match). Pass `all` to skip this filter.
      schema:
        type: string
        example: "HBO,AMC"
    limit:
      name: limit
      in: query
      description: Number of results per page (default 20).
      schema:
        type: integer
        minimum: 1
        maximum: 900
        default: 20
    page:
      name: page
      in: query
      description: Page number, 1-based (default 1).
      schema:
        type: integer
        minimum: 1
        default: 1
  schemas:
...
    RatedEpisode:
      type: object
      properties:
        season:
          type: integer
        episode:
          type: integer
        title:
          type: string
        description:
          type: string
        id:
          type: string
          description: IMDb episode ID (e.g. tt1267295).
        url:
          type: string
          format: uri
        release_date:
          type: string
          format: date
        users_rating:
          type: number
          description: IMDb user rating 0–10.
        users_rating_count:
          type: integer
...
    PaginatedRatedEpisodes:
      type: object
      properties:
        page:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/RatedEpisodeWithShow"
        total_pages:
          type: integer
        total_results:
          type: integer
    RatedEpisodeWithShow:
      allOf:
        - $ref: "#/components/schemas/RatedEpisode"
        - type: object
          properties:
            tvshow:
              type: object
              properties:
                id:
                  type: integer
                item_type:
                  type: string
                title:
                  type: string
                image:
                  type: string
                  format: uri
                networks:
                  type: object
                seasons_number:
                  type: integer
                status:
                  type: string
                  enum: [Canceled, Ended, Ongoing, Pilot, Unknown]
    Error:
      type: object
      properties:
        message:
          type: string
          example: "No matching items found."
  responses:
    RateLimitExceeded:
      description: Rate limit exceeded. Retry after the indicated delay.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /episodes/rated:
    get:
      operationId: getRatedEpisodes
      summary: Get highest or lowest rated episodes across all TV shows
      description: |
        Returns IMDb-rated episodes ranked across all TV shows. Supports episode-level
        sorting (by rating, ascending or descending) and TV show–level filters (genre,
        platform, network, status, and more). Each result includes the parent show's
        metadata alongside the episode data.

        Only episodes with at least `minimum_users_rating_count` user ratings (default 100)
        are included to filter out low-confidence ratings.
      tags: [Episodes]
      parameters:
        - name: title
          in: query
          description: Search across the TV show title, episode title, or TV show networks (partial match, case-insensitive).
          schema:
            type: string
        - $ref: "#/components/parameters/genres"
        - $ref: "#/components/parameters/platforms"
        - $ref: "#/components/parameters/networks"
        - name: is_active
          in: query
          description: |
            `true` returns only active TV shows, `false` returns only inactive TV shows,
            `true,false` (or omitting this parameter) returns both.
          schema:
            type: string
            example: "true"
        - name: is_adult
          in: query
          description: |
            `false` (default) returns only non-adult TV shows, `true` returns only
            adult-flagged TV shows, `true,false` returns both.
          schema:
            type: string
            default: "false"
            example: "true"
        - name: must_see
          in: query
          description: |
            `true` returns only TV shows with Metacritic "Must-See" badge, `false`
            returns only those without it, `true,false` returns both.
          schema:
            type: string
            example: "true"
        - name: users_certified
          in: query
          description: |
            `true` returns only TV shows with Rotten Tomatoes "Verified Audience"
            certification, `false` returns only uncertified shows, `true,false` returns both.
          schema:
            type: string
            example: "true"
        - name: critics_certified
          in: query
          description: |
            `true` returns only TV shows with Rotten Tomatoes "Certified Fresh" status,
            `false` returns only non-certified shows, `true,false` returns both.
          schema:
            type: string
            example: "true"
        - name: directors
          in: query
          description: Comma-separated director names (partial match). Pass `all` to skip this filter.
          schema:
            type: string
        - name: production_companies
          in: query
          description: Comma-separated production company names (partial match). Pass `all` to skip this filter.
          schema:
            type: string
        - name: status
          in: query
          description: Filter the parent TV show by production status.
          schema:
            type: string
            enum: [canceled, ended, ongoing, pilot, unknown]
        - name: filtered_seasons
          in: query
          description: |
            Comma-separated season numbers to restrict results to.
            Example: `1,2` returns only episodes from seasons 1 and 2.
          schema:
            type: string
            example: "1,2"
        - name: minimum_ratings
          in: query
          description: Minimum IMDb user rating (0–10 scale).
          schema:
            type: number
            minimum: 0
            maximum: 10
            example: 9.0
        - name: minimum_users_rating_count
          in: query
          description: Minimum number of IMDb user ratings required per episode (default 100).
          schema:
            type: integer
            minimum: 0
            default: 100
        - $ref: "#/components/parameters/releaseDateRange"
        - name: order
          in: query
          description: "`desc` returns top-rated episodes, `asc` returns lowest-rated episodes (default: `desc`)."
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/page"
      responses:
        "200":
          description: Paginated rated episodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRatedEpisodes"
        "400":
          description: Invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No matching items found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method not allowed.
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.